Add my git command for pruning a remote repository.
When you run this command, specifying one of your
remotes to prune, every branch on the  remote repository
that is not also on your local repository will be deleted.

This is a great command, for the rare times when you need it,
such as when working with a personal downstream branch, that's
meant to follow the state of your own personal branch.